2006-02-24  Jim Garrison  <garrison@users.sourceforge.net>
	* backend/sound_handler_sdl.cpp: fixed int, unsigned int comparisions
	* libbase/grid_index.h: fixed int, unsigned int comparisions
	* libbase/image_filters.cpp: fixed int, unsigned int comparisions
	* libbase/triangulate_impl.h: fixed int, unsigned int comparisions
	* libgeometry/kd_tree_dynamic.cpp: fixed int, unsigned int comparisions
	* libgeometry/tqt.cpp: fixed int, unsigned int comparisions
	* plugin/plugin.cpp: fixed int, unsigned int comparisions
	* server/Movie.cpp: fixed int, unsigned int comparisions
	* server/Sprite.cpp: fixed int, unsigned int comparisions
	* server/action.cpp: fixed int, unsigned int comparisions
	* server/button.cpp: fixed int, unsigned int comparisions
	* server/dlist.cpp: fixed int, unsigned int comparisions
	* server/font.cpp: fixed int, unsigned int comparisions
	* server/fontlib.cpp: fixed int, unsigned int comparisions
	* server/impl.cpp: fixed int, unsigned int comparisions
	* server/morph2.cpp: fixed int, unsigned int comparisions
	* server/shape.cpp: fixed int, unsigned int comparisions
	* server/styles.cpp: fixed int, unsigned int comparisions
	* server/tesselate.cpp: fixed int, unsigned int comparisions
	* server/text.cpp: fixed int, unsigned int comparisions
	* server/xmlnode.cpp: fixed int, unsigned int comparisions
	* server/xmlsocket.cpp: fixed int, unsigned int comparisions

+++ backend/sound_handler_sdl.cpp
@@ -170,7 +170,7 @@ struct SDL_sound_handler : gnash::sound_handler
 	virtual void	play_sound(int sound_handle, int loop_count /* other params */)
 	// Play the index'd sample.
 	{
-		if (sound_handle >= 0 && sound_handle < m_samples.size())
+	  if (sound_handle >= 0 && sound_handle < (int) m_samples.size())
 		{
 			if (m_samples[sound_handle])
 			{
@@ -183,7 +183,7 @@ struct SDL_sound_handler : gnash::sound_handler
 	
 	virtual void	stop_sound(int sound_handle)
 	{
-		if (sound_handle < 0 || sound_handle >= m_samples.size())
+	  if (sound_handle < 0 || sound_handle >= (int) m_samples.size())
 		{
 			// Invalid handle.
 			return;
@@ -205,7 +205,7 @@ struct SDL_sound_handler : gnash::sound_handler
 	virtual void	delete_sound(int sound_handle)
 	// this gets called this when it's done with a sample.
 	{
-		if (sound_handle >= 0 && sound_handle < m_samples.size())
+	  if (sound_handle >= 0 && sound_handle < (int) m_samples.size())
 		{
 			Mix_Chunk*	chunk = m_samples[sound_handle];
 			if (chunk)
